+++ test/integration/test-batch-callback.js
@@ -459,6 +459,7 @@ describe('batch callback', function () {
       if (err) return done(err);
       conn.query('DROP TABLE IF EXISTS nonRewritableBatch');
       conn.query('CREATE TABLE nonRewritableBatch(id int, t varchar(256))');
+      conn.beginTransaction();
       conn.batch(
         'INSERT INTO nonRewritableBatch(id, t) VALUES (?,?)',
         [
@@ -481,8 +482,10 @@ describe('batch callback', function () {
                   t: 'jack'
                 }
               ]);
-              conn.end();
-              done();
+              conn.commit((err, res) => {
+                conn.end();
+                done();
+              });
             });
           }
         }
@@ -730,34 +733,36 @@ describe('batch callback', function () {
 
       conn.query('DROP TABLE IF EXISTS nonRewritableHoldersErr');
       conn.query('CREATE TABLE nonRewritableHoldersErr(id int, t varchar(256))');
-      conn.batch(
-        'INSERT INTO nonRewritableHoldersErr(id, t) VALUES (:id2,:id1)',
-        [
-          { id2: 1, id1: 'john' },
-          { id1: 'jack', id2: 2 }
-        ],
-        (err, res) => {
-          if (err) {
-            conn.end();
-            done(err);
-          } else {
-            conn.query('SELECT * FROM nonRewritableHoldersErr', (err, res) => {
-              assert.deepEqual(res, [
-                {
-                  id: 1,
-                  t: 'john'
-                },
-                {
-                  id: 2,
-                  t: 'jack'
-                }
-              ]);
+      conn.beginTransaction(() => {
+        conn.batch(
+          'INSERT INTO nonRewritableHoldersErr(id, t) VALUES (:id2,:id1)',
+          [
+            { id2: 1, id1: 'john' },
+            { id1: 'jack', id2: 2 }
+          ],
+          (err, res) => {
+            if (err) {
               conn.end();
-              done();
-            });
+              done(err);
+            } else {
+              conn.query('SELECT * FROM nonRewritableHoldersErr', (err, res) => {
+                assert.deepEqual(res, [
+                  {
+                    id: 1,
+                    t: 'john'
+                  },
+                  {
+                    id: 2,
+                    t: 'jack'
+                  }
+                ]);
+                conn.end();
+                done();
+              });
+            }
           }
-        }
-      );
+        );
+      });
     });
   };
 
@@ -843,7 +848,7 @@ describe('batch callback', function () {
         (err) => {
           if (!err) {
             conn.end();
-            done(new Error('must have thrown error !'));
+            return done(new Error('must have thrown error !'));
           }
           assert.isTrue(err != null);
           if (!isXpand()) {
@@ -897,6 +902,7 @@ describe('batch callback', function () {
       const conn = await base.createConnection({ compress: useCompression, bulk: true });
       conn.query('DROP TABLE IF EXISTS blabla');
       conn.query('CREATE TABLE blabla(i int, i2 int)');
+      conn.beginTransaction();
       await conn.batch('INSERT INTO `blabla` values (?, ?)', [
         [1, 2],
         [1, undefined]
@@ -907,6 +913,7 @@ describe('batch callback', function () {
         { i: 1, i2: null }
       ]);
       conn.query('DROP TABLE IF EXISTS blabla');
+      conn.commit();
       conn.end();
     });
 
@@ -1090,29 +1097,31 @@ describe('batch callback', function () {
       const conn = base.createCallbackConnection({ compress: useCompression, bulk: true });
       conn.query('DROP TABLE IF EXISTS blabla');
       conn.query('CREATE TABLE blabla(i int, i2 int)');
-      conn.batch(
-        'INSERT INTO `blabla` values (?,?)',
-        [
-          [1, 2],
-          [1, undefined]
-        ],
-        (err, res) => {
-          conn.query('SELECT * from blabla', (err, rows) => {
-            if (err) {
-              done(err);
-            } else {
-              assert.deepEqual(rows, [
-                { i: 1, i2: 2 },
-                { i: 1, i2: null }
-              ]);
-              conn.query('DROP TABLE IF EXISTS blabla', (err) => {
-                conn.end();
-                done();
-              });
-            }
-          });
-        }
-      );
+      conn.beginTransaction(() => {
+        conn.batch(
+          'INSERT INTO `blabla` values (?,?)',
+          [
+            [1, 2],
+            [1, undefined]
+          ],
+          (err, res) => {
+            conn.query('SELECT * from blabla', (err, rows) => {
+              if (err) {
+                done(err);
+              } else {
+                assert.deepEqual(rows, [
+                  { i: 1, i2: 2 },
+                  { i: 1, i2: null }
+                ]);
+                conn.query('DROP TABLE IF EXISTS blabla', (err) => {
+                  conn.end();
+                  done();
+                });
+              }
+            });
+          }
+        );
+      });
     });
 
     it('simple batch offset date', function (done) {