+++ test/integration/test-connection-opts.js
@@ -8,6 +8,7 @@ const { assert } = require('chai');
 
 describe('connection option', () => {
   it('with undefined collation', function (done) {
+    if (process.env.srv === 'xpand') this.skip();
     base
       .createConnection({ charset: 'unknown' })
       .then(() => {
@@ -17,6 +18,15 @@ describe('connection option', () => {
         assert(err.message.includes('Unknown'));
         done();
       });
+  });
+
+  it('collation with no id', async function () {
+    if (process.env.srv === 'xpand' || process.env.srv === 'mariadb-es' || process.env.srv === 'mariadb-es-test')
+      this.skip();
+    if (!shareConn.info.isMariaDB() || !shareConn.info.hasMinVersion(11, 2, 0)) this.skip();
+    const conn = await base.createConnection();
+    await conn.query('set NAMES utf8mb4 COLLATE uca1400_vietnamese_ai_ci');
+    conn.end();
   });
 
   it('wrong IANA timezone', function (done) {
@@ -35,12 +45,12 @@ describe('connection option', () => {
   });
 
   it('ensure Etc/GMT', async function () {
-    let conn = await base.createConnection({ timezone: 'Etc/GMT-8', debug: true });
+    let conn = await base.createConnection({ timezone: 'Etc/GMT-8' });
     let res = await conn.query('SELECT @@time_zone as t');
     assert.equal(res[0].t, '+08:00');
     conn.end();
 
-    conn = await base.createConnection({ timezone: 'GMT-8', debug: true });
+    conn = await base.createConnection({ timezone: 'GMT-8' });
     res = await conn.query('SELECT @@time_zone as t');
     assert.equal(res[0].t, '-08:00');
     conn.end();
@@ -72,7 +82,7 @@ describe('connection option', () => {
   });
 
   it('timezone +0h', async function () {
-    const conn = await base.createConnection({ timezone: '+00:00', debug: true });
+    const conn = await base.createConnection({ timezone: '+00:00' });
 
     let d = new Date('2000-01-01T00:00:00Z');
     let res = await conn.query("SELECT UNIX_TIMESTAMP('2000-01-01T00:00:00') tt", [d]);