+++ test/integration/test-debug.js
@@ -54,8 +54,8 @@ describe('debug', () => {
   });
 
   //ensure that debug from previous test are written to console
-  afterEach(() => {
-    logger.close();
+  afterEach(async function () {
+    await closeLogger(logger);
     try {
       fs.unlinkSync(tmpLogFile);
     } catch (e) {}
@@ -213,7 +213,7 @@ describe('debug', () => {
                 .then(() => {
                   const serverVersion = conn.serverVersion();
                   const data = fs.readFileSync(tmpLogFile, 'utf8');
-                  let range = [9500, 12000];
+                  let range = [8900, 12000];
                   assert(
                     data.length > range[0] && data.length < range[1],
                     'wrong data length : ' +
@@ -261,7 +261,7 @@ describe('debug', () => {
       const conn = await base.createConnection({ debug: true });
       const res = await conn.query("SELECT '1'");
       conn.end();
-      const range = [3700, 5800];
+      const range = [3600, 5800];
       assert(
         data.length > range[0] && data.length < range[1],
         'wrong data length : ' +
@@ -330,6 +330,7 @@ describe('debug', () => {
   }
 
   it('fast path command debug', async function () {
+    if (isXpand()) this.skip();
     await testPingDebug(false);
   });
 
@@ -357,7 +358,7 @@ describe('debug', () => {
     const serverVersion = conn.serverVersion();
     if (process.env.srv === 'maxscale' || process.env.srv === 'skysql' || process.env.srv === 'skysql-ha')
       compress = false;
-    const range = compress ? [90, 180] : [90, 170];
+    const range = compress ? [60, 180] : [60, 170];
     const data = fs.readFileSync(tmpLogFile, 'utf8');
     assert.isTrue(data.includes('PING'));
     assert.isTrue(data.includes('QUIT'));
@@ -378,3 +379,28 @@ describe('debug', () => {
     );
   }
 });
+
+const closeLogger = async function (logger) {
+  const promises = [];
+
+  // close all transports -- transports dont use promises...
+  // syslog close function emits 'closed' when done
+  // daily-rotate-file close function emits 'finish' when done
+  for (const transport of logger.transports) {
+    if (transport.close) {
+      const promise = new Promise((resolve) => {
+        transport.once('closed', () => {
+          resolve();
+        });
+        transport.once('finish', () => {
+          resolve();
+        });
+      });
+      promises.push(promise);
+      // transport.close();  <-- invoked by logger.close()
+    }
+  }
+
+  logger.close();
+  return Promise.all(promises);
+};