+++ test/integration/test-pool-callback.js
@@ -361,7 +361,7 @@ describe('Pool callback', () => {
       assert.equal(pool.escapeId('good_$one'), '`good_$one`');
       assert.equal(pool.escape(''), "''");
       assert.equal(pool.escapeId('f:a'), '`f:a`');
-      assert.equal(pool.escapeId('`f:a`'), '`f:a`');
+      assert.equal(pool.escapeId('`f:a`'), '```f:a```');
       assert.equal(pool.escapeId('good_`è`one'), '`good_``è``one`');
       pool.end();
       pool2.end();