+++ test/integration/test-pool.js
@@ -354,7 +354,7 @@ describe('Pool', () => {
       assert.equal(pool.escapeId('good_$one'), '`good_$one`');
       assert.equal(pool.escape(''), "''");
       assert.equal(pool.escapeId('f:a'), '`f:a`');
-      assert.equal(pool.escapeId('`f:a`'), '`f:a`');
+      assert.equal(pool.escapeId('`f:a`'), '```f:a```');
       assert.equal(pool.escapeId('good_`è`one'), '`good_``è``one`');
       await pool.end();
       await pool2.end();
@@ -839,7 +839,13 @@ describe('Pool', () => {
   });
 
   it('pool getConnection timeout', function (done) {
-    if (process.env.srv === 'maxscale' || process.env.srv === 'skysql' || process.env.srv === 'skysql-ha') this.skip();
+    if (
+      process.env.srv === 'maxscale' ||
+      process.env.srv === 'skysql' ||
+      process.env.srv === 'skysql-ha' ||
+      process.env.srv === 'xpand'
+    )
+      this.skip();
     const pool = base.createPool({ connectionLimit: 1, acquireTimeout: 200 });
     let errorThrown = false;
     pool
@@ -1106,7 +1112,7 @@ describe('Pool', () => {
             assert.equal(pool.totalConnections(), 10);
             assert.equal(pool.idleConnections(), 0);
             assert.isOk(pool.taskQueueSize() > 8000);
-          }, 100);
+          }, 200);
         } else {
           assert.equal(pool.activeConnections(), 10);
           assert.equal(pool.totalConnections(), 10);